requirements.txt
azure_storage==0.37.0
Flask==3.0.3
openai==1.42.0
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
